refactored controllers to use common worker in `util/worker/worker.go` and harmonized constants.
