+++ pkg/util/worker/worker.go
@@ -1,0 +1,62 @@
+package worker
+
+import (
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/client-go/util/workqueue"
+	"k8s.io/klog/v2"
+	"sync"
+	"time"
+)
+
+const (
+	DefaultMaxRetries = 15
+)
+
+// Run creates and runs a worker routine that just processes items in the
+// specified queue. The worker will run until stopCh is closed. The worker will be
+// added to the wait group when started and marked done when finished.
+func Run(queue workqueue.RateLimitingInterface, resourceType string, maxRetries int, forgetAfterSuccess bool, reconciler func(key string) error, stopCh <-chan struct{}, waitGroup *sync.WaitGroup) {
+	waitGroup.Add(1)
+	go func() {
+		wait.Until(worker(queue, resourceType, maxRetries, forgetAfterSuccess, reconciler), time.Second, stopCh)
+		waitGroup.Done()
+	}()
+}
+
+// worker runs a worker thread that just dequeues items, processes them, and marks them done.
+// If reconciler returns an error, requeue the item up to maxRetries before giving up.
+// It enforces that the reconciler is never invoked concurrently with the same key.
+// If forgetAfterSuccess is true, it will cause the queue to forget the item should reconciliation
+// have no error.
+func worker(queue workqueue.RateLimitingInterface, resourceType string, maxRetries int, forgetAfterSuccess bool, reconciler func(key string) error) func() {
+	return func() {
+		exit := false
+		for !exit {
+			exit = func() bool {
+				key, quit := queue.Get()
+				if quit {
+					return true
+				}
+				defer queue.Done(key)
+
+				err := reconciler(key.(string))
+				if err == nil {
+					if forgetAfterSuccess {
+						queue.Forget(key)
+					}
+					return false
+				}
+
+				if queue.NumRequeues(key) < maxRetries {
+					klog.V(4).Infof("Error syncing %s %v: %v", resourceType, key, err)
+					queue.AddRateLimited(key)
+					return false
+				}
+
+				klog.V(4).Infof("Dropping %s %q out of the queue: %v", resourceType, key, err)
+				queue.Forget(key)
+				return false
+			}()
+		}
+	}
+}
